requirements.txt
aiogram>=2.0,<3.0
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv